requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
mistralai>=0.1.0
PyPDF2>=3.0.0
